# gpt_action_openapi.yaml
openapi: 3.1.0
info:
  title: Python Drive Organizer API
  version: 1.7-MVP01.1
  description: >-
    Read-only operations access a SQLite index of Google Drive file and folder
    metadata, including paged Tree viewing and authenticated TXT/DOCX/XLSX
    export. Contacts operations access the SQLite address-book snapshot. Plain
    email has a read-only preview and confirmed
    text-only send with no Drive calls. The legacy confirmed operation sends
    one exact Drive blob as an attachment. Enhanced email uses a read-only
    preview followed by a confirmed send of 1-5 attachments or Drive links.
    Link send may create only non-discoverable anyone/reader permissions after
    warning that a forwarded link remains accessible. No operation creates,
    updates, moves, copies, trashes, or deletes Drive items.
servers:
  - url: https://drive-api.sungwony.pe.kr
    description: Cloudflare HTTPS endpoint for Python Drive Organizer
security:
  - BearerAuth: []
paths:
...
  /contacts/search:
    post:
      operationId: searchContacts
      summary: Search the SQLite contact directory
      description: >-
        Search current SQLite contacts by free text or exact organization,
        name, title, or email filters. Use this for name-based recipients.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSearchRequest"
      responses:
        "200":
          description: Ranked matching contacts from the SQLite snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSearchResponse"
        "400":
          description: At least one non-empty search criterion is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /contacts/status:
    get:
      operationId: getContactsStatus
      summary: Get Contacts synchronization status
      description: >-
        Return the latest Contacts sync status, last successful completion
        time, and row statistics from SQLite. This does not refresh Contacts.
      x-openai-isConsequential: false
      responses:
        "200":
          description: Latest Contacts synchronization state and statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key configured separately in the GPT Action authentication settings.
  responses:
    Unauthorized:
      description: Bearer authentication is missing or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: A query or path parameter failed validation.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - $ref: "#/components/schemas/HTTPValidationError"
    ServiceUnavailable:
      description: Authentication configuration or the SQLite index is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ContactSearchRequest:
      type: object
      additionalProperties: false
      description: >-
        Supply q or at least one exact filter. q and filters are combined with
        AND. Blank criteria are rejected by the API.
      properties:
        q:
          type:
            - string
            - "null"
          minLength: 1
          description: Text matched against normalized name, organization, title, and email.
        organization:
          type:
            - string
            - "null"
          minLength: 1
          description: Optional normalized exact organization filter.
        name:
          type:
            - string
            - "null"
          minLength: 1
          description: Optional normalized exact name filter.
        title:
          type:
            - string
            - "null"
          minLength: 1
          description: Optional normalized exact title filter.
        email:
          type:
            - string
            - "null"
          minLength: 1
          description: Optional case-insensitive exact email filter.
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Maximum number of contacts to return.
    ContactItem:
      type: object
      additionalProperties: false
      required:
        - contact_id
        - organization
        - name
        - title
        - email
        - phone
        - email_usable
        - conflict_code
      properties:
        contact_id:
          type: string
          description: Exact opaque ID to preserve and re-read before preview.
        organization:
          type:
            - string
            - "null"
        name:
          type: string
        title:
          type:
            - string
            - "null"
        email:
          type:
            - string
            - "null"
        phone:
          type:
            - string
            - "null"
        email_usable:
          type: boolean
          description: False means this contact must not be used for email delivery.
        conflict_code:
          type:
            - string
            - "null"
          description: A non-null conflict blocks automatic email use.
    ContactSearchResponse:
      type: object
      additionalProperties: false
      required: [total, showing, items]
      properties:
        total:
          type: integer
          minimum: 0
        showing:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContactItem"
    ContactsStatusResponse:
      type: object
      additionalProperties: false
      required:
        - latest_sync_id
        - latest_sync_status
        - last_success_at
        - rows_seen
        - valid_rows
        - inserted
        - updated
        - deleted
        - unchanged
        - invalid
        - conflicts
      properties:
        latest_sync_id:
          type:
            - string
            - "null"
        latest_sync_status:
          type:
            - string
            - "null"
        last_success_at:
          type:
            - string
            - "null"
        rows_seen:
          type: integer
          minimum: 0
        valid_rows:
          type: integer
          minimum: 0
        inserted:
          type: integer
          minimum: 0
        updated:
          type: integer
          minimum: 0
        deleted:
          type: integer
          minimum: 0
        unchanged:
          type: integer
          minimum: 0
        invalid:
          type: integer
          minimum: 0
        conflicts:
          type: integer
          minimum: 0
    ContactErrorResponse:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorItem"
    ValidationErrorItem:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string